% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{all_games}
\alias{all_games}
\title{returns all games and data currently in the steamspy database}
\usage{
all_games()
}
\value{
A tibble with all games and data currently in the Steamspy database
}
\description{
returns all games and data currently in the steamspy database
}
