% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{game_detail}
\alias{game_detail}
\title{Return more detailed information on one specific game, given appid
currently need to figure out how to parse the json, has a list column
of tags !!!WARNING!!! Unlike most of the other functions, this could
conceivably run up against the API limits, which is 4 calls / second
if you just blindly run it in a loop. Don't be a jerk!}
\usage{
game_detail(appid = 730)
}
\arguments{
\item{appid}{An integer corresponding to the steam game, look up via
the summary top... or other ways}
}
\description{
Return more detailed information on one specific game, given appid
currently need to figure out how to parse the json, has a list column
of tags !!!WARNING!!! Unlike most of the other functions, this could
conceivably run up against the API limits, which is 4 calls / second
if you just blindly run it in a loop. Don't be a jerk!
}
