% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{genre_games}
\alias{genre_games}
\title{Return a tibble with games corresponding to a specific genre.
Replace spaces with +}
\usage{
genre_games(genre)
}
\arguments{
\item{genre}{a string}
}
\value{
A tibble with games corresponding to a specific genre.
}
\description{
Return a tibble with games corresponding to a specific genre.
Replace spaces with +
}
