% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{parse_sspy}
\alias{parse_sspy}
\title{parse steam spy API json}
\usage{
parse_sspy(json)
}
\arguments{
\item{json}{A json produced by the Steam Spy API}
}
\value{
A tibble of steam spy values, usually ~ 100 games, unless
allgames() was usedue
}
\description{
parse steam spy API json
}
