% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{tag_games}
\alias{tag_games}
\title{Return a tibble with games corresponding to a specific tag.
Replace spaces with +}
\usage{
tag_games(tag)
}
\arguments{
\item{tag}{a string}
}
\value{
A tibble with games corresponding to a specific tag.
}
\description{
Return a tibble with games corresponding to a specific tag.
Replace spaces with +
}
