% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{top_2wk}
\alias{top_2wk}
\title{return a tibble with the top 100 games from the last 2 weeks.}
\usage{
top_2wk()
}
\value{
A tibble with the top 100 games from the last 2 weeks.
}
\description{
return a tibble with the top 100 games from the last 2 weeks.
}
