% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_steamspy.R
\name{top_forever}
\alias{top_forever}
\title{return a tibble with the top 100 games since March 2009.}
\usage{
top_forever()
}
\value{
A tibble with the top 100 games since March 2009.
}
\description{
return a tibble with the top 100 games since March 2009.
}
